+++ packages/date-range-input/src/components/date-range-input/Component.tsx
@@ -319,7 +319,7 @@ export const DateRangeInput = React.forwardRef<HTMLInputElement, DateRangeInputP
         const handleInputKeyDown = (event: KeyboardEvent<HTMLInputElement>) => {
             const isCopy = (event.metaKey || event.ctrlKey) && event.key === 'c';
 
-            if (disableUserInput && !isCopy) {
+            if (disableUserInput && !isCopy && event.key !== 'Tab') {
                 event.preventDefault();
             }
         };
@@ -414,9 +414,7 @@ export const DateRangeInput = React.forwardRef<HTMLInputElement, DateRangeInputP
         };
 
         const handleCalendarChange = (date?: number) => {
-            if (date) {
-                updatePeriod(date);
-            }
+            updatePeriod(date);
         };
 
         const handleCalendarWrapperMouseDown = (event: MouseEvent<HTMLDivElement>) => {