+++ packages/tag/src/Component.mobile.tsx
@@ -9,9 +9,9 @@ import styles from './mobile.module.css';
 const colorStyles = {
     default: defaultColors,
     inverted: invertedColors,
-}
+};
 
-export type TagMobileProps = Omit<BaseTagProps, 'styles'>;
+export type TagMobileProps = Omit<BaseTagProps, 'styles' | 'colorStylesMap'>;
 
 export const TagMobile = forwardRef<HTMLButtonElement, TagMobileProps>((restProps, ref) => (
     <BaseTag {...restProps} colorStylesMap={colorStyles} ref={ref} styles={styles} />