+++ packages/tooltip/src/Component.mobile.tsx
@@ -17,31 +17,31 @@ export const TooltipMobile: React.FC<TooltipMobileProps> = ({
     targetClassName,
     children,
     getPortalContainer,
+    targetTag: TargetTag = 'div',
     ...restProps
 }) => (
     <Fragment>
         <BottomSheet
-            {...restProps}
-            container={getPortalContainer}
-            onClose={onClose}
             actionButton={
                 <ButtonMobile view='secondary' block={true} size='s' onClick={onClose}>
                     {actionButtonTitle}
                 </ButtonMobile>
             }
+            {...restProps}
+            container={getPortalContainer}
+            onClose={onClose}
         >
             {content}
         </BottomSheet>
 
         {/** TODO: проверить тултип на доступность */}
-        {/* eslint-disable-next-line jsx-a11y/click-events-have-key-events, jsx-a11y/no-static-element-interactions */}
-        <div
+        <TargetTag
             ref={targetRef as React.Ref<HTMLDivElement>}
             onClick={onOpen}
             className={cn(styles.target, targetClassName)}
         >
             {children?.props.disabled && <div className={styles.overlap} />}
             {children}
-        </div>
+        </TargetTag>
     </Fragment>
 );