+++ packages/calendar-input/CHANGELOG.md
@@ -1,5 +1,28 @@
 # Change Log
 
+## 9.0.4
+
+### Patch Changes
+
+### [#815](https://github.com/core-ds/core-components/pull/815)
+
+-   Исправлена проблема, из-за которой при включенной пропсе disableUserInput невозможно было переставить фокус клавишей Tab
+
+## 9.0.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   calendar@7.0.3
+
+## 9.0.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   date-input@4.3.2
+    -   calendar@7.0.2
+
 ## 9.0.1
 
 ### Patch Changes