+++ packages/calendar-range/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Change Log
 
+## 7.2.4
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   calendar-input@9.0.4
+
+## 7.2.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   calendar@7.0.3
+    -   calendar-input@9.0.3
+
+## 7.2.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   date-input@4.3.2
+    -   calendar@7.0.2
+    -   calendar-input@9.0.2
+
 ## 7.2.1
 
 ### Patch Changes