+++ packages/calendar-with-skeleton/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 5.1.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   calendar@7.0.3
+
+## 5.1.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   calendar@7.0.2
+
 ## 5.1.1
 
 ### Patch Changes