+++ packages/calendar/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 7.0.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   modal@8.0.3
+
+## 7.0.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@9.0.2
+    -   icon-button@6.2.2
+    -   modal@8.0.2
+
 ## 7.0.1
 
 ### Patch Changes