+++ packages/circular-progress-bar/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 3.1.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   typography@4.1.0
+
 ## 3.1.0
 
 ### Minor Changes