+++ packages/custom-picker-button/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @alfalab/core-components-custom-picker-button
 
+## 2.0.4
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   select@15.0.4
+    -   picker-button@11.0.4
+
+## 2.0.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   select@15.0.3
+    -   picker-button@11.0.3
+
+## 2.0.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   picker-button@11.0.2
+    -   select@15.0.2
+    -   custom-button@3.3.2
+
 ## 2.0.1
 
 ### Patch Changes