+++ packages/gallery/CHANGELOG.md
@@ -1,5 +1,35 @@
 # Change Log
 
+## 5.3.5
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   tooltip@7.1.0
+
+## 5.3.4
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   tooltip@7.0.4
+
+## 5.3.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   typography@4.1.0
+    -   tooltip@7.0.3
+
+## 5.3.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   icon-button@6.2.2
+    -   tooltip@7.0.2
+
 ## 5.3.1
 
 ### Patch Changes