+++ packages/notification/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## 6.2.3
+
+### Patch Changes
+
+### [#812](https://github.com/core-ds/core-components/pull/812)
+
+-   fix: increased css specificity
+
+## 6.2.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   toast-plate@6.0.2
+
 ## 6.2.1
 
 ### Patch Changes