+++ packages/phone-input/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## 7.2.3
+
+### Patch Changes
+
+### [#802](https://github.com/core-ds/core-components/pull/802)
+
+-   Исправлена ошибка, из-за которой в поле вставлялась лишняя 7 при вводе значения перед +7
+
+## 7.2.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   masked-input@6.2.2
+
 ## 7.2.1
 
 ### Patch Changes