+++ packages/select/CHANGELOG.md
@@ -1,5 +1,31 @@
 # Change Log
 
+## 15.0.4
+
+### Patch Changes
+
+### [#821](https://github.com/core-ds/core-components/pull/821)
+
+-   Исправлено ошибка, из-за которой невозможно было выбрать опцию в мобильном селекте
+
+## 15.0.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   bottom-sheet@6.7.0
+    -   modal@8.0.3
+
+## 15.0.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@12.0.2
+    -   button@9.0.2
+    -   bottom-sheet@6.6.2
+    -   modal@8.0.2
+
 ## 15.0.1
 
 ### Patch Changes