+++ packages/slider-input/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 7.2.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@12.0.2
+
 ## 7.2.1
 
 ### Patch Changes